% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_misc.R
\name{availableModels}
\alias{availableModels}
\title{Available models}
\usage{
availableModels()
}
\value{
a data.frame containing models implemented in MixtComp
\describe{
\item{model}{model name}
\item{data.type}{data type}
\item{format}{Special format required for individuals}
\item{missing.formats}{accepted formats (separated by a ;) for missing values}
\item{hyperparameter}{Required hyperparameters in the paramStr elements of model object}
\item{comments}{comments about the model}
\item{reference}{link to article}
}
}
\description{
Get information about models implemented in MixtComp
}
\seealso{
\code{mixtCompLearn}
}
\author{
Quentin Grimonprez
}
