% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_formatParameter.R
\name{completeAlgo}
\alias{completeAlgo}
\title{Add the missing element to algo parameter

Add the missing element to algo parameter with default values}
\usage{
completeAlgo(algo)
}
\arguments{
\item{algo}{a list with the different algo parameters for rmc function}
}
\value{
algo parameter with all required elements (see \link{createAlgo} function)
}
\description{
Add the missing element to algo parameter

Add the missing element to algo parameter with default values
}
\author{
Quentin Grimonprez
}
