% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_formatFunctions.R
\name{convertFunctionalToVector}
\alias{convertFunctionalToVector}
\title{Convert a MixtComp functional string into a list of 2 vectors}
\usage{
convertFunctionalToVector(x)
}
\arguments{
\item{x}{a string containing a functional observation (cf example)}
}
\value{
a list of 2 vectors: \code{time} and \code{value}
}
\description{
Convert a MixtComp functional string into a list of 2 vectors
}
\examples{
convertFunctionalToVector("1:5,1.5:12,1.999:2.9")

}
\author{
Quentin Grimonprez
}
