% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_formatParameter.R
\name{formatData}
\alias{formatData}
\title{Format the data parameter required by rmc}
\usage{
formatData(data)
}
\arguments{
\item{data}{data parameter as data.frame, matrix or list}
}
\value{
data as a list of characters
}
\description{
format data.frame or matrix in list of character
}
\author{
Quentin Grimonprez
}
