% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_formatParameter.R
\name{formatModel}
\alias{formatModel}
\title{Format the model parameter}
\usage{
formatModel(model)
}
\arguments{
\item{model}{description of model used per variable}
}
\value{
model as a list where each element is the model applied to a variable (list with elements type and paramStr)
}
\description{
Format the model list for rmc/rmcMultiRun functions:
- add paramStr when missing
- ensure the list format of each element
}
\author{
Quentin Grimonprez
}
