% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_heatmap.R
\name{heatmapVar}
\alias{heatmapVar}
\title{Heatmap of the similarities between variables about clustering}
\usage{
heatmapVar(output, pkg = c("ggplot2", "plotly"), ...)
}
\arguments{
\item{output}{object returned by \emph{mixtCompLearn} function from \emph{RMixtComp} or \emph{rmcMultiRun} function
from \emph{RMixtCompIO}}

\item{pkg}{"ggplot2" or "plotly". Package used to plot}

\item{...}{arguments to be passed to plot_ly. For pkg = "ggplot2", addValues = TRUE prints similarity values on the heatmap}
}
\description{
Heatmap of the similarities between variables about clustering
}
\details{
The similarities between variables j and h is defined by Delta(j,h)
\deqn{Delta(j,h) = 1 - \sqrt{(1/n) * \sum_{i=1}^n \sum_{k=1}^K (P(Z_i=k|x_{ij}) - P(Z_i=k|x_{ih}))^2}}
}
\examples{
if (requireNamespace("RMixtCompIO", quietly = TRUE)) {
  dataLearn <- list(
    var1 = as.character(c(rnorm(50, -2, 0.8), rnorm(50, 2, 0.8))),
    var2 = as.character(c(rnorm(50, 2), rpois(50, 8)))
  )

  model <- list(
    var1 = list(type = "Gaussian", paramStr = ""),
    var2 = list(type = "Poisson", paramStr = "")
  )

  algo <- list(
    nClass = 2,
    nInd = 100,
    nbBurnInIter = 100,
    nbIter = 100,
    nbGibbsBurnInIter = 100,
    nbGibbsIter = 100,
    nInitPerClass = 3,
    nSemTry = 20,
    confidenceLevel = 0.95,
    ratioStableCriterion = 0.95,
    nStableCriterion = 10,
    mode = "learn"
  )

  resLearn <- RMixtCompIO::rmcMultiRun(algo, dataLearn, model, nRun = 3)

  # plot
  heatmapVar(resLearn)
}

}
\seealso{
\code{\link{computeSimilarityVar}}

Other plot: 
\code{\link{heatmapClass}()},
\code{\link{heatmapTikSorted}()},
\code{\link{histMisclassif}()},
\code{\link{plot.MixtComp}()},
\code{\link{plotConvergence}()},
\code{\link{plotDataBoxplot}()},
\code{\link{plotDataCI}()},
\code{\link{plotDiscrimClass}()},
\code{\link{plotDiscrimVar}()},
\code{\link{plotParamConvergence}()},
\code{\link{plotProportion}()}
}
\author{
Matthieu MARBAC
}
\concept{plot}
