% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_formatFunctions.R
\name{refactorCategorical}
\alias{refactorCategorical}
\title{Rename a categorical value}
\usage{
refactorCategorical(
  data,
  oldCateg = unique(data),
  newCateg = 1:length(oldCateg)
)
}
\arguments{
\item{data}{matrix/data.frame/vector containing the data}

\item{oldCateg}{vector containing categories to change}

\item{newCateg}{vector containing new categorical values}
}
\value{
Data with new categorical values
}
\description{
Rename a categorical value
}
\examples{
dat <- c("single", "married", "married", "divorced", "single")
refactorCategorical(dat, c("single", "married", "divorced"), 1:3)

}
\author{
Quentin Grimonprez
}
