% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_formatFunctions.R
\name{convertFunctionalToVector}
\alias{convertFunctionalToVector}
\title{Convert a mixtcomp string into 2-columns matrix}
\usage{
convertFunctionalToVector(x)
}
\arguments{
\item{x}{a string containing a fonctional observation (cf example)}
}
\value{
matrix
}
\description{
Convert a mixtcomp string into 2-columns matrix
}
\examples{
convertFunctionalToVector("1:5,1.5:12,1.999:2.9")

}
\author{
Quentin Grimonprez
}
