% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_heatmap.R
\name{heatmapVar}
\alias{heatmapVar}
\title{Heatmap of the similarities between variables about clustering}
\usage{
heatmapVar(output, pkg = c("ggplot2", "plotly"), ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompLearn}}

\item{pkg}{"ggplot2" or "plotly". Package used to plot}

\item{...}{arguments to be passed to plot_ly. For pkg = "ggplot2", addValues = TRUE prints similarity values on the heatmap}
}
\description{
Heatmap of the similarities between variables about clustering
}
\details{
The similarities between variables j and h is defined by Delta(j,h)
\deqn{Delta(j,h) = 1 - \sqrt{(1/n) * \sum_{i=1}^n \sum_{k=1}^K (P(Z_i=k|x_{ij}) - P(Z_i=k|x_{ih}))^2}}
}
\examples{
\donttest{
data(simData)
 
# define the algorithm's parameters
algo <- createAlgo()

# keep only 3 variables
model <- simData$model$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtComp in unsupervised clustering mode + data as matrix
res <- mixtCompLearn(simData$dataLearn$matrix, model, algo, nClass = 2:4)

# plot
heatmapVar(res)
heatmapVar(res$res[[2]])
}

}
\seealso{
\code{\link{computeSimilarityVar}}

Other plot: \code{\link{heatmapClass}},
  \code{\link{heatmapTikSorted}},
  \code{\link{histMisclassif}},
  \code{\link{plot.MixtComp}},
  \code{\link{plotConvergence}}, \code{\link{plotCrit}},
  \code{\link{plotDataBoxplot}}, \code{\link{plotDataCI}},
  \code{\link{plotDiscrimClass}},
  \code{\link{plotDiscrimVar}},
  \code{\link{plotParamConvergence}},
  \code{\link{plotProportion}}
}
\author{
Matthieu MARBAC
}
\concept{plot}
