% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_histAndLine.R
\name{histMisclassif}
\alias{histMisclassif}
\title{Histogram of the misclassification probabilities}
\usage{
histMisclassif(output, pkg = c("ggplot2", "plotly"), ...)
}
\arguments{
\item{output}{object returned by \emph{mixtCompLearn} function from \emph{RMixtComp} or \emph{rmcMultiRun} function from
\emph{RMixtCompIO}}

\item{pkg}{"ggplot2" or "plotly". Package used to plot}

\item{...}{arguments to be passed to plot_ly}
}
\description{
Histogram of the misclassification probabilities
}
\details{
Missclassification probability of observation i is denoted err_i
err_i = 1 - max_{k=1,\ldots,K} P(Z_i=k|x_i)
Histograms of err_i's can be plotted for a specific class, all classes or every class
}
\examples{
if (requireNamespace("RMixtCompIO", quietly = TRUE)) {
  dataLearn <- list(
    var1 = as.character(c(rnorm(50, -2, 0.8), rnorm(50, 2, 0.8))),
    var2 = as.character(c(rnorm(50, 2), rpois(50, 8)))
  )

  model <- list(
    var1 = list(type = "Gaussian", paramStr = ""),
    var2 = list(type = "Poisson", paramStr = "")
  )

  algo <- list(
    nClass = 2,
    nInd = 100,
    nbBurnInIter = 100,
    nbIter = 100,
    nbGibbsBurnInIter = 100,
    nbGibbsIter = 100,
    nInitPerClass = 3,
    nSemTry = 20,
    confidenceLevel = 0.95,
    ratioStableCriterion = 0.95,
    nStableCriterion = 10,
    mode = "learn"
  )

  resLearn <- RMixtCompIO::rmcMultiRun(algo, dataLearn, model, nRun = 3)

  # plot
  histMisclassif(resLearn)
}

}
\seealso{
Other plot: 
\code{\link{heatmapClass}()},
\code{\link{heatmapTikSorted}()},
\code{\link{heatmapVar}()},
\code{\link{plot.MixtComp}()},
\code{\link{plotConvergence}()},
\code{\link{plotDataBoxplot}()},
\code{\link{plotDataCI}()},
\code{\link{plotDiscrimClass}()},
\code{\link{plotDiscrimVar}()},
\code{\link{plotParamConvergence}()},
\code{\link{plotProportion}()}
}
\author{
Matthieu MARBAC
}
\concept{plot}
