% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_barplot.R
\name{plotDiscrimVar}
\alias{plotDiscrimVar}
\title{Barplot of the discriminative power of the variables}
\usage{
plotDiscrimVar(output, ylim = c(0, 1), pkg = c("ggplot2", "plotly"),
  ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompLearn}}

\item{ylim}{vector of length 2 defining the range of y-axis}

\item{pkg}{"ggplot2" or "plotly". Package used to plot}

\item{...}{arguments to be passed to plot_ly}
}
\description{
Barplot of the discriminative power of the variables
}
\details{
The discriminative power of variable j is defined by 1 - C(j)
\deqn{C(j)=  -\sum_{k=1}^K \sum_{i=1}^n P(Z_i=k|x_{ij}) ln(P(Z_i=k|x_{ij})) / (n*\log(K))}
}
\examples{
require(RMixtCompIO) # for learning a mixture model
dataLearn <- list(var1 = as.character(c(rnorm(50, -2, 0.8), rnorm(50, 2, 0.8))),
                  var2 = as.character(c(rnorm(50, 2), rpois(50, 8))))
                  
model <- list(var1 = list(type = "Gaussian", paramStr = ""),
              var2 = list(type = "Poisson", paramStr = ""))

algo <- list(
  nClass = 2,
  nInd = 100,
  nbBurnInIter = 100,
  nbIter = 100,
  nbGibbsBurnInIter = 100,
  nbGibbsIter = 100,
  nInitPerClass = 3,
  nSemTry = 20,
  confidenceLevel = 0.95,
  ratioStableCriterion = 0.95,
  nStableCriterion = 10,
  mode = "learn"
)

resLearn <- rmcMultiRun(algo, dataLearn, model, nRun = 3)

# plot
plotDiscrimVar(resLearn)

}
\seealso{
\code{\link{computeDiscrimPowerVar}}

Other plot: \code{\link{heatmapClass}},
  \code{\link{heatmapTikSorted}}, \code{\link{heatmapVar}},
  \code{\link{histMisclassif}},
  \code{\link{plot.MixtComp}},
  \code{\link{plotConvergence}}, \code{\link{plotCrit}},
  \code{\link{plotDataBoxplot}}, \code{\link{plotDataCI}},
  \code{\link{plotDiscrimClass}},
  \code{\link{plotParamConvergence}},
  \code{\link{plotProportion}}
}
\author{
Matthieu MARBAC
}
\concept{plot}
